requirements.txt
langchain_groq 
langchain_core
python-dotenv
streamlit
langchain_community
pypdf
faiss-cpu
groq
sentence-transformers
langchain-huggingface